fix bugs from sql

+++ db/roles.sql
@@ -1,19 +1,34 @@
 -- depends_on: versions.sql
 
+DO LANGUAGE plpgsql $wrapper$
+BEGIN
+
+  -- Drop roles if exists cascade
+  IF EXISTS (SELECT 1 FROM pg_roles WHERE rolname='comicstor_postgraphql') THEN
+    DROP OWNED BY comicstor_postgraphql CASCADE;
+    DROP ROLE comicstor_postgraphql;
+  END IF;
+
+  IF EXISTS (SELECT 1 FROM pg_roles WHERE rolname='comicstor_anonymous') THEN
+    DROP OWNED BY comicstor_anonymous CASCADE;
+    DROP ROLE comicstor_anonymous;
+  END IF;
+
+  IF EXISTS (SELECT 1 FROM pg_roles WHERE rolname='comicstor_user') THEN
+    DROP OWNED BY comicstor_user CASCADE;
+    DROP ROLE comicstor_user;
+  END IF;
+
+END;
+$wrapper$;
+
 -- Internal user for connecting from api to db
-DROP OWNED BY comicstor_postgraphql CASCADE;
-DROP ROLE IF EXISTS comicstor_postgraphql;
 CREATE ROLE comicstor_postgraphql LOGIN PASSWORD 'xyz';
 
 -- Anonymous user role
-DROP OWNED BY comicstor_anonymous CASCADE;
-DROP ROLE IF EXISTS comicstor_anonymous;
 CREATE ROLE comicstor_anonymous;
+GRANT comicstor_anonymous TO comicstor_postgraphql;
 
 -- Authenticated user role
-DROP OWNED BY comicstor_user CASCADE;
-DROP ROLE IF EXISTS comicstor_user;
 CREATE ROLE comicstor_user;
-
-GRANT comicstor_anonymous TO comicstor_postgraphql;
 GRANT comicstor_user TO comicstor_postgraphql;
